refactors handleClick to reduce repeated code, uses useEffect to handle conditional rerendering
