+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Game from './components/GameHook.js';
+import Game from './components/Game.js';
 import './index.css'
 
 ReactDOM.render(
